requirements.txt
torch>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0
numpy>=1.24.0
transformers>=4.36.0
peft>=0.10.0
pillow>=9.0.0